even

+++ 16-binary_tree_is_perfect.c
@@ -54,7 +54,7 @@ if (!tree)
 return (1);
 left_height = binary_tree_height(tree->left);
 right_height = binary_tree_height(tree->right);
-return ((int)(left_height == right_height) &&
+return ((left_height == right_height) &&
 	binary_tree_is_perfect(tree->left) &&
 	binary_tree_is_perfect(tree->right));
 }
